third_party/blink/web_tests/platform/win7/virtual/prerender/external/wpt/speculation-rules/prerender/restriction-window-move-expected.txt

This is a testharness.js-based test.
FAIL a prerendering page cannot move its window by executing moveTo. assert_equals: expected "PASS" but got "FAIL: Error: assert_equals: y position for primary expected 1 but got 0"
PASS a prerendering page cannot move its window by executing moveBy.
Harness: the test ran to completion.
